Move Router component and make other minor cleanup changes
