+++ src/component/Router.js
@@ -1,29 +1,25 @@
 // This page uses the React Router library to make links to navigate from page to page
 import React from 'react';
-import './App.css';
-import Home from './component/Home';
-import FetchItems from './component/FetchItems';
-import AddItem from './component/AddItem';
+import '../App.css';
+import Home from './Home';
+import FetchItems from './FetchItems';
+import AddItem from './AddItem';
 import { BrowserRouter as Router, Route, Link, Switch } from 'react-router-dom'
 
 function RouterComponent() {
-    // console.log('is this the router page?')
     return (
         <Router>
-            {/* Our links. So far, we have three, though 
-            that might change as the project grows */}
-            <Link to="/">Home</Link>
             <Link to="/list">List</Link>
             <Link to="/add">Add</Link>
             <Switch>
-                <Route exact path="/">
-                    <Home />
-                </Route>
                 <Route exact path="/list">
                     <FetchItems />
                 </Route>
                 <Route exact path="/add">
                     <AddItem />
+                </Route>
+                <Route exact path="/">
+                    <Home />
                 </Route>
             </Switch>
         </Router>
